Rename Embed page to Blank page
