+++ src/router.ts
@@ -3,7 +3,7 @@ import Router from "vue-router";
 import BaseLayout from "./layouts/default.vue";
 import Main from "./views/Main.vue";
 import Welcome from "./views/Home.vue";
-import Embed from "./views/Embed.vue";
+import Blank from "./views/Blank.vue";
 
 Vue.use(Router);
 
@@ -30,7 +30,7 @@ export default new Router({
     {
       path: "/blank",
       name: "blank",
-      component: Embed,
+      component: Blank,
     },
   ],
 });
