+++ tslint.json
@@ -15,6 +15,7 @@
     "ordered-imports": false,
     "object-literal-sort-keys": false,
     "no-consecutive-blank-lines": false,
+    "no-console": [false],
     "max-line-length": [
       true,
       {
